+++ crates/ruma-appservice-api/src/thirdparty/get_protocol.rs
@@ -7,11 +7,14 @@ pub mod v1 {
     //!
     //! [spec]: https://spec.matrix.org/latest/application-service-api/#get_matrixappv1thirdpartyprotocolprotocol
 
+    use std::collections::BTreeMap;
+
     use ruma_common::{
         api::{request, response, Metadata},
         metadata,
-        thirdparty::Protocol,
+        thirdparty::{Protocol, ProtocolInstance, ProtocolInstanceInit},
     };
+    use serde::{Deserialize, Serialize};
 
     const METADATA: Metadata = metadata! {
         method: GET,
@@ -35,7 +38,7 @@ pub mod v1 {
     pub struct Response {
         /// Metadata about the protocol.
         #[ruma_api(body)]
-        pub protocol: Protocol,
+        pub protocol: AppserviceProtocol,
     }
 
     impl Request {
@@ -47,8 +50,55 @@ pub mod v1 {
 
     impl Response {
         /// Creates a new `Response` with the given protocol.
-        pub fn new(protocol: Protocol) -> Self {
+        pub fn new(protocol: AppserviceProtocol) -> Self {
             Self { protocol }
+        }
+    }
+
+    /// Metadata about a third party protocol, as returned by an appservice to a homeserver.
+    ///
+    /// To create an instance of this type, first create a [`ProtocolInit`] and convert it via
+    /// `AppserviceProtocol::from` / `.into()`.
+    ///
+    /// [`ProtocolInit`]: ruma_common::thirdparty::ProtocolInit
+    pub type AppserviceProtocol = Protocol<AppserviceProtocolInstance>;
+
+    /// Metadata about an instance of a third party protocol, as returned by an appservice to a
+    /// homeserver.
+    ///
+    /// To create an instance of this type, first create a [`ProtocolInstanceInit`] and convert it
+    /// via `AppserviceProtocolInstance::from` / `.into()`.
+    #[derive(Clone, Debug, Deserialize, Serialize)]
+    #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+    pub struct AppserviceProtocolInstance {
+        /// A human-readable description for the protocol, such as the name.
+        pub desc: String,
+
+        /// An optional content URI representing the protocol.
+        #[serde(skip_serializing_if = "Option::is_none")]
+        pub icon: Option<String>,
+
+        /// Preset values for `fields` the client may use to search by.
+        pub fields: BTreeMap<String, String>,
+
+        /// A unique identifier across all instances.
+        pub network_id: String,
+    }
+
+    impl From<ProtocolInstanceInit> for AppserviceProtocolInstance {
+        fn from(init: ProtocolInstanceInit) -> Self {
+            let ProtocolInstanceInit { desc, fields, network_id } = init;
+            Self { desc, icon: None, fields, network_id }
+        }
+    }
+
+    impl From<AppserviceProtocolInstance> for ProtocolInstance {
+        fn from(value: AppserviceProtocolInstance) -> Self {
+            let AppserviceProtocolInstance { desc, icon, fields, network_id } = value;
+            let mut instance =
+                ProtocolInstance::from(ProtocolInstanceInit { desc, fields, network_id });
+            instance.icon = icon;
+            instance
         }
     }
 }