+++ crates/ruma-common/src/thirdparty.rs
@@ -15,11 +15,11 @@ use crate::{
 
 /// Metadata about a third party protocol.
 ///
-/// To create an instance of this type, first create a `ProtocolInit` and convert it via
+/// To create an instance of this type, first create a [`ProtocolInit`] and convert it via
 /// `Protocol::from` / `.into()`.
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-pub struct Protocol {
+pub struct Protocol<I = ProtocolInstance> {
     /// Fields which may be used to identify a third party user.
     pub user_fields: Vec<String>,
 
@@ -37,16 +37,30 @@ pub struct Protocol {
     pub field_types: BTreeMap<String, FieldType>,
 
     /// A list of objects representing independent instances of configuration.
-    pub instances: Vec<ProtocolInstance>,
+    pub instances: Vec<I>,
 }
 
-/// Initial set of fields of `Protocol`.
+impl<I> Protocol<I> {
+    /// Convert this `Protocol<I>` to a `Protocol<J>`.
+    pub fn into<J: From<I>>(self) -> Protocol<J> {
+        let Self { user_fields, location_fields, icon, field_types, instances } = self;
+        Protocol {
+            user_fields,
+            location_fields,
+            icon,
+            field_types,
+            instances: instances.into_iter().map(J::from).collect(),
+        }
+    }
+}
+
+/// Initial set of fields of [`Protocol`].
 ///
-/// This struct will not be updated even if additional fields are added to `Prococol` in a new
-/// (non-breaking) release of the Matrix specification.
+/// This struct will not be updated even if additional fields are added to [`Protocol`] in
+/// a new (non-breaking) release of the Matrix specification.
 #[derive(Debug)]
 #[allow(clippy::exhaustive_structs)]
-pub struct ProtocolInit {
+pub struct ProtocolInit<I = ProtocolInstance> {
     /// Fields which may be used to identify a third party user.
     pub user_fields: Vec<String>,
 
@@ -60,20 +74,20 @@ pub struct ProtocolInit {
     pub field_types: BTreeMap<String, FieldType>,
 
     /// A list of objects representing independent instances of configuration.
-    pub instances: Vec<ProtocolInstance>,
+    pub instances: Vec<I>,
 }
 
-impl From<ProtocolInit> for Protocol {
-    fn from(init: ProtocolInit) -> Self {
+impl<I> From<ProtocolInit<I>> for Protocol<I> {
+    fn from(init: ProtocolInit<I>) -> Self {
         let ProtocolInit { user_fields, location_fields, icon, field_types, instances } = init;
         Self { user_fields, location_fields, icon, field_types, instances }
     }
 }
 
-/// Metadata about an instance of a third party protocol.
+/// Metadata about an instance of a third party protocol, as returned by a homeserver to a client.
 ///
-/// To create an instance of this type, first create a `ProtocolInstanceInit` and convert it via
-/// `ProtocolInstance::from` / `.into()`.
+/// To create an instance of this type, first create a [`ProtocolInstanceInit`] or an
+/// `AppserviceProtocolInstance` and convert it via `ProtocolInstance::from` / `.into()`.
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
 pub struct ProtocolInstance {
@@ -90,18 +104,21 @@ pub struct ProtocolInstance {
     /// A unique identifier across all instances.
     pub network_id: String,
 
-    /// A unique identifier across all instances.
+    /// A unique identifier for this instance on the homeserver.
     ///
-    /// See [matrix-spec#833](https://github.com/matrix-org/matrix-spec/issues/833).
-    #[cfg(feature = "unstable-unspecified")]
+    /// This is a field added by the homeserver to `AppserviceProtocolInstance`. It can be used as
+    /// the value of [`RoomNetwork::ThirdParty`] in a request to the `get_public_rooms_filtered`
+    /// endpoint.
+    ///
+    /// [`RoomNetwork::ThirdParty`]: crate::directory::RoomNetwork::ThirdParty
     #[serde(skip_serializing_if = "Option::is_none")]
     pub instance_id: Option<String>,
 }
 
-/// Initial set of fields of `Protocol`.
+/// Initial set of fields of [`ProtocolInstance`].
 ///
-/// This struct will not be updated even if additional fields are added to `Prococol` in a new
-/// (non-breaking) release of the Matrix specification.
+/// This struct will not be updated even if additional fields are added to [`ProtocolInstance`] in a
+/// new (non-breaking) release of the Matrix specification.
 #[derive(Debug)]
 #[allow(clippy::exhaustive_structs)]
 pub struct ProtocolInstanceInit {
@@ -118,14 +135,7 @@ pub struct ProtocolInstanceInit {
 impl From<ProtocolInstanceInit> for ProtocolInstance {
     fn from(init: ProtocolInstanceInit) -> Self {
         let ProtocolInstanceInit { desc, fields, network_id } = init;
-        Self {
-            desc,
-            icon: None,
-            fields,
-            network_id,
-            #[cfg(feature = "unstable-unspecified")]
-            instance_id: None,
-        }
+        Self { desc, icon: None, fields, network_id, instance_id: None }
     }
 }