+++ crates/ruma-common/CHANGELOG.md
@@ -19,11 +19,14 @@ Breaking changes:
 - `MatrixVersion::(into/from)_parts` are no longer exposed as public methods.
   They were usually used to sort `MatrixVersion`s, now the `PartialOrd` and
   `Ord` implementations can be used instead.
+- `Protocol` and `ProtocolInit` are generic on the protocol instance type.
 
 Improvements:
 
 - `MatrixVersion` implements `PartialOrd` and `Ord`. The variants are ordered by
   release date, with a newer version being greater than an older version.
+- `ProtocolInstance` has an `instance_id` field, due to a clarification in the
+  spec.
 
 # 0.15.1